# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: BizzHive Marketplace API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /products:
    get:
      operationId: listProducts
      tags: [products]
      summary: List digital products
      parameters:
        - name: categoryId
          in: query
          schema:
            type: integer
        - name: vendorId
          in: query
          schema:
            type: integer
        - name: search
          in: query
          schema:
            type: string
        - name: type
          in: query
          schema:
            type: string
            enum: [ebook, template, software, asset, other]
        - name: featured
          in: query
          schema:
            type: boolean
        - name: sortBy
          in: query
          schema:
            type: string
            enum: [newest, popular, price_low, price_high, rating]
      responses:
        "200":
          description: List of products
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Product"
    post:
      operationId: createProduct
      tags: [products]
      summary: Create a digital product
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateProductBody"
      responses:
        "201":
          description: Product created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
  /products/{id}:
    get:
      operationId: getProduct
      tags: [products]
      summary: Get product details
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Product details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
        "404":
          description: Product not found
    patch:
      operationId: updateProduct
      tags: [products]
      summary: Update a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateProductBody"
      responses:
        "200":
          description: Product updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Product"
    delete:
      operationId: deleteProduct
      tags: [products]
      summary: Delete a product
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Product deleted
...
  /vendors/{vendorId}/payout-settings:
    put:
      operationId: updateVendorPayoutSettings
      tags: [payouts]
      summary: Update vendor payout settings (MoMo / bank)
      parameters:
        - name: vendorId
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/VendorPayoutSettingsBody"
      responses:
        "200":
          description: Payout settings updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/VendorPayoutSettings"
  /payouts/process/{orderId}:
    post:
      operationId: processOrderPayouts
      tags: [payouts]
      summary: Process vendor payouts for a completed order
      parameters:
        - name: orderId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Payout results
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PayoutResult"
...
components:
  schemas:
...
    Product:
      type: object
      properties:
        id:
          type: integer
        title:
          type: string
        slug:
          type: string
        description:
          type: ["string", "null"]
        thumbnail:
          type: ["string", "null"]
        price:
          type: number
          format: float
        currency:
          type: string
        productType:
          type: string
          enum: [ebook, template, software, asset, audio, other]
        fileUrl:
          type: ["string", "null"]
        previewUrl:
          type: ["string", "null"]
        licenseTerms:
          type: ["string", "null"]
        rating:
          type: number
          format: float
        reviewsCount:
          type: integer
        salesCount:
          type: integer
        featured:
          type: boolean
        published:
          type: boolean
        vendorId:
          type: integer
        vendorName:
          type: string
        vendorAvatar:
          type: ["string", "null"]
        categoryId:
          type: integer
        categoryName:
          type: string
        createdAt:
          type: string
          format: date-time
      required:
        [
          id,
          title,
          slug,
          price,
          currency,
          productType,
          rating,
          reviewsCount,
          salesCount,
          featured,
          published,
          vendorId,
          vendorName,
          categoryId,
          categoryName,
          createdAt,
        ]
    CreateProductBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        price:
          type: number
        currency:
          type: string
        productType:
          type: string
          enum: [ebook, template, software, asset, audio, other]
        fileUrl:
          type: string
        previewUrl:
          type: string
          description: Streamable preview clip — used for audio/beat listings.
        licenseTerms:
          type: string
          description: Licensing terms the buyer agrees to (audio/beat listings).
        vendorId:
          type: integer
        categoryId:
          type: integer
      required: [title, price, vendorId, categoryId, productType]
    UpdateProductBody:
      type: object
      properties:
        title:
          type: string
        description:
          type: string
        thumbnail:
          type: string
        price:
          type: number
        currency:
          type: string
        productType:
          type: string
          enum: [ebook, template, software, asset, audio, other]
        fileUrl:
          type: string
        previewUrl:
          type: string
          description: Streamable preview clip — used for audio/beat listings.
        licenseTerms:
          type: string
          description: Licensing terms the buyer agrees to (audio/beat listings).
        categoryId:
          type: integer
        published:
          type: boolean
        featured:
          type: boolean
...
    VendorPayoutSettingsBody:
      type: object
      properties:
        momoNumber:
          type: string
        momoNetwork:
          type: string
          enum: [MTN, Vodafone, AirtelTigo]
        email:
          type: string
      required: [momoNumber, momoNetwork]
    VendorPayoutSettings:
      type: object
      properties:
        vendorId:
          type: integer
        momoNumber:
          type: ["string", "null"]
        momoNetwork:
          type: ["string", "null"]
        email:
          type: ["string", "null"]
        payoutPercentage:
          type: number
      required: [vendorId, payoutPercentage]
...
    PayoutResult:
      type: object
      properties:
        orderId:
          type: integer
        payouts:
          type: array
          items:
            type: object
            properties:
              vendorId:
                type: integer
              vendorName:
                type: string
              amount:
                type: number
              status:
                type: string
              reference:
                type: string
            required: [vendorId, vendorName, amount, status]
      required: [orderId, payouts]

    # ------------------------------------------------------------------
    # Bookable session slots
    # ------------------------------------------------------------------
...